# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Plataforma.dev API
  description: |
    Complete API documentation for Plataforma.dev - the AI-first enterprise platform.
    
    This API provides comprehensive endpoints for:
    - User authentication and authorization
    - Module management and deployment
    - Database operations with real-time updates
    - File storage and management
    - Real-time communication via WebSockets
    
    ## Authentication
    
    The API uses JWT (JSON Web Tokens) for authentication. Include the token in the Authorization header:
    
    ```
    Authorization: Bearer <your-jwt-token>
    ```
    
    ## Rate Limiting
    
    API requests are rate-limited to ensure fair usage:
    - **General API**: 1000 requests per hour per IP
    - **Authentication**: 10 attempts per 15 minutes per IP
    - **File uploads**: 100 requests per hour per user
    
    Rate limit headers are included in responses:
    - `X-RateLimit-Limit`: Request limit per time window
    - `X-RateLimit-Remaining`: Remaining requests in current window
    - `X-RateLimit-Reset`: Time when the rate limit resets
    
    ## Error Handling
    
    The API uses conventional HTTP response codes and returns detailed error information:
    
    ```json
    {
      "success": false,
      "error": {
        "code": "VALIDATION_FAILED",
        "message": "Request validation failed",
        "details": {
          "field": "email",
          "reason": "Invalid email format"
        }
      },
...
servers:
...
security:
  - BearerAuth: []
paths:
...
  /auth/login:
    post:
      tags:
        - Authentication
      summary: User login
      description: Authenticate user with email and password
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            examples:
              email_login:
                summary: Email login
                value:
                  email: "user@example.com"
                  password: "password123"
                  rememberMe: true
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
...
  /auth/logout:
    post:
      tags:
        - Authentication
      summary: User logout
      description: Invalidate current session and token
      responses:
        '200':
          description: Logout successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogoutResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: user@example.com
        password:
          type: string
          format: password
          example: password123
        rememberMe:
          type: boolean
          default: false
    LoginResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        token:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        refreshToken:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        user:
          $ref: '#/components/schemas/User'
        expiresAt:
          type: string
          format: date-time
          example: 2024-08-26T18:30:00Z
...
    LogoutResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Logged out successfully
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 123e4567-e89b-12d3-a456-426614174000
        email:
          type: string
          format: email
          example: user@example.com
        name:
          type: string
          example: John Doe
        avatar:
          type: string
          format: uri
          nullable: true
        roles:
          type: array
          items:
            type: string
          example: [user, manager]
        permissions:
          type: array
          items:
            type: string
        preferences:
          type: object
          additionalProperties: true
        created_at:
          type: string
          format: date-time
        last_login:
          type: string
          format: date-time
          nullable: true
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              example: VALIDATION_FAILED
            message:
              type: string
              example: Request validation failed
            details:
              type: object
              additionalProperties: true
        timestamp:
          type: string
          format: date-time
  responses:
    BadRequest:
      description: Bad request - Invalid input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: VALIDATION_FAILED
              message: Request validation failed
              details:
                field: email
                reason: Invalid email format
            timestamp: 2024-08-26T10:30:00Z
    Unauthorized:
      description: Unauthorized - Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: UNAUTHORIZED
              message: Authentication required
            timestamp: 2024-08-26T10:30:00Z
...
    RateLimitExceeded:
      description: Too many requests
      headers:
        X-RateLimit-Limit:
          description: Request limit per time window
          schema:
            type: integer
        X-RateLimit-Remaining:
          description: Remaining requests in current window
          schema:
            type: integer
        X-RateLimit-Reset:
          description: Time when the rate limit resets
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: RATE_LIMIT_EXCEEDED
              message: Too many requests, try again later
              details:
                limit: 1000
                window: 3600
                retry_after: 1800
            timestamp: 2024-08-26T10:30:00Z
...